Dynamic Leadership (#60)
Updates cron to use fully dyanmic leadership, supporting scaling up and
down without manual intervention.

The stored leadership proto has been updated to include a UID field to
ensure a leadership key is only owned by a single replica of that ID.

The PatitionID and PartitionTotal options have been removed, in favour
of a user provided ID string which can be any non-empty string value to
identify that cron replica instance.

Leadership is now dynamic in that it will attempt to elect itself with a
partition total which it discovers from the leadership key space. Upon
leadership, it will continue to wait for quorum for all other
leaders in the cluster. On leadership events (crons being added or
removed from the cluster), leadrship will be re-elected by observing the
new leadership total, and updating oneself. Uid is used to prevent a
replica with the same ID assuming leadership on the same key.

The partition modulo index is calculated by the position in the sorted
list or leadership key strings. This means that the consumer need not
supply a partition total nor ID uint, and can simply provide a string
which is unique to that replica.

Naturally supports active-passive for replicas with the same ID.

---

Adds support for a `WatchLeadership` channel which will receive the
replica data of leaders in the cluster, both on initial election, and on
leadership partition rebalancing.

+++ internal/leadership/partitioner/modulo_test.go
@@ -15,9 +15,9 @@ func Test_parter(t *testing.T) {
 	t.Parallel()
 
 	tests := []struct {
-		id          uint32
-		total       uint32
-		partitionID uint32
+		id          uint64
+		total       uint64
+		partitionID uint64
 		exp         bool
 	}{
 		{
